添加对pi发送函数

+++ Core/Src/scheduler.c
@@ -53,7 +53,7 @@ static void Loop_20Hz(void) //50ms执行一次
 static void Loop_10Hz(void) //100ms执行一次
 {
     printf("D=%.2f\r\n",DISTANCE);
-    send_to_pi(DISTANCE);
+    send_to_pi((int)DISTANCE);
 }
 
 static void Loop_5Hz(void) //200ms执行一次
